Wrong incrementation end and data types limit
The code works well but because the first two values are fixed, the program return for example with 3: `1, 1, 2, 3` so the values up to the fourth one. An other problem is with values above the 47th, the wanted value is 2,971,215,073 but because `int` go from -2,147,483,648 to 2,147,483,647, the value overflows and you have instead -1,323,752,223. Using `unsigned long long int`, you can go from 0 to 18,446,744,073,709,551,615. That way, it works for more numbers.